server/requirements.txt
certifi==2018.4.16
pandas==0.23.4
django_postgres_copy==2.3.7
Django==2.2.1
djangorestframework==3.9.4
django-cors-headers==2.4.1
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.3
setuptools==39.1.0